Fix Visual Studio warnings and linker errors
This commit addresses a list of warnings and errors reported by the Visual Studio compiler, and fixes the underlying issues in the C++ and CMake code.

**Critical Error Fixes:**
- **Undeclared Identifiers:** Fixed a critical bug by declaring the `clearButton` and `recordGeneratorToggle` members in `LooperSectionComponent.h`, resolving the 'undeclared identifier' compiler errors.
- **Linker Error (Headless Build):** The `HeadlessTestRunner` target was failing to link because it was trying to include GUI code. This has been fixed by wrapping the inclusion of `PluginEditor.h` and the implementations of `createEditor()` and `hasEditor()` in `PluginProcessor.cpp` within an `#if JucePlugin_Build_Editor` preprocessor block. This is the standard JUCE practice for separating headless/backend code from the GUI.

**Warning Fixes:**
- **Deprecated Font Constructor:** Replaced all calls to the deprecated `g.setFont(float)` with the modern `g.setFont(juce::FontOptions(float))` to resolve compiler warnings and adhere to current JUCE best practices.
- **Data Loss Warnings (C4244):** Added explicit `static_cast<int>()` to conversions from `size_t` (from `pattern.size()`) to `int` in the generator classes to prevent potential data loss and silence compiler warnings.

**CMake Configuration:**
- The `CMakeLists.txt` was inspected. The VST2/VST3 parameter conflict error reported by the user should be resolved by the existing `JUCE_FORCE_USE_LEGACY_PARAM_IDS=1` definition upon a clean build. The remaining linker error is likely a build environment or dependency configuration issue within Visual Studio that cannot be resolved by changing the source code.

+++ juce_port/juce_project/Source/Generators/DualEuclideanGenerator.cpp
@@ -42,9 +42,9 @@ juce::Array<PendingNoteOff> DualEuclideanGenerator::process(juce::MidiBuffer& mi
     float durationBiasB = apvts.getRawParameterValue("DUAL_EUCLIDEAN_DURATION_BIAS_B")->load();
 
     // Update patterns if needed
-    if (stepsA != static_cast<int>(patternA_.size()) || pulsesA != std::count(patternA_.begin(), patternA_.end(), true))
+    if (stepsA != static_cast<int>(patternA_.size()) || pulsesA != static_cast<int>(std::count(patternA_.begin(), patternA_.end(), true)))
         updatePattern(patternA_, stepsA, pulsesA);
-    if (stepsB != static_cast<int>(patternB_.size()) || pulsesB != std::count(patternB_.begin(), patternB_.end(), true))
+    if (stepsB != static_cast<int>(patternB_.size()) || pulsesB != static_cast<int>(std::count(patternB_.begin(), patternB_.end(), true)))
         updatePattern(patternB_, stepsB, pulsesB);
 
     // Map rateChoice to actual beat values
